requirements.txt
anyio==4.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
numpy==1.26.3
pandas==2.2.0
pillow==10.2.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-telegram-bot==20.7
pytz==2024.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0
